[witness] fix function_enter AVOID handling and extract witness functions
For real FUNCTION_CALL instructions, replace pc++ with assume(false) when a
function_enter AVOID waypoint matches: execution has reached the avoided call
site, so the path is inconsistent with the witness and must be killed.

Merge symex_witness_inlined_function_enter into symex_witness_function_enter
(returns bool: true = path killed) so both real calls and inlined nondet calls
share one implementation. Extract all witness-steering functions
(symex_witness_function_return, symex_witness_function_enter,
symex_witness_branching, symex_witness_assert) into
builtin_functions/witness.cpp.
