+++ src/goto-symex/symex_main.cpp
@@ -127,29 +127,7 @@ void goto_symext::claim(const expr2tc &claim_expr, const std::string &msg)
     check_incremental(new_expr, msg))
     return; // Verification succeeded, no further action needed
 
-  if (validate_witness && !has_prefix(msg, "unwinding assertion loop"))
-  {
-    const size_t seg = cur_state->cur_seg;
-    const waypoint *target_wp = nullptr;
-    if (seg < cur_state->witness_segs.size())
-      for (const auto &wp : cur_state->witness_segs[seg])
-        if (wp.type == waypoint::target)
-        {
-          target_wp = &wp;
-          break;
-        }
-
-    if (
-      !target_wp ||
-      cur_state->source.pc->location.get_line() != target_wp->line_id ||
-      cur_state->witness_target_reached)
-      new_expr = gen_true_expr();
-    else
-    {
-      cur_state->advance_witness_position();
-      cur_state->witness_target_reached = true;
-    }
-  }
+  symex_witness_assert(new_expr, msg);
 
   // add assertion to the target equation
   assertion(new_expr, msg);
@@ -273,74 +251,6 @@ goto_symext::symex_resultt goto_symext::get_symex_result()
     target, total_claims, remaining_claims, simplified_claims);
 }
 
-static void substitute_result(expr2tc &e, const expr2tc &ret_val)
-{
-  if (is_symbol2t(e) && to_symbol2t(e).thename == "\\result")
-  {
-    e = ret_val;
-    return;
-  }
-  if (is_constant_int2t(e))
-  {
-    if (
-      is_unsignedbv_type(ret_val->type) &&
-      to_constant_int2t(e).value.is_negative())
-      log_warning(
-        "witness: function_return constraint compares signed constant {} "
-        "against unsigned return type; constraint may be trivially false",
-        integer2string(to_constant_int2t(e).value));
-    e = from_integer(to_constant_int2t(e).value, ret_val->type);
-    return;
-  }
-  e->Foreach_operand([&](expr2tc &op) { substitute_result(op, ret_val); });
-}
-
-void goto_symext::symex_witness_function_return(
-  expr2tc ret_val,
-  const irep_idt &call_line)
-{
-  if (cur_state->cur_seg >= cur_state->witness_segs.size())
-    return;
-
-  const auto &seg = cur_state->witness_segs[cur_state->cur_seg];
-  for (const waypoint &wp : seg)
-  {
-    if (wp.type == waypoint::function_enter)
-    {
-      if (wp.line_id.empty() || wp.line_id != call_line)
-        continue;
-      if (wp.action == waypoint::avoid)
-      {
-        cur_state->source.pc++;
-        return;
-      }
-
-      cur_state->advance_witness_position();
-      return;
-    }
-
-    if (wp.type != waypoint::function_return)
-      continue;
-    if (!wp.parsed_cond.valid)
-      continue;
-    if (wp.line_id != call_line)
-      continue;
-
-    cur_state->rename(ret_val);
-
-    expr2tc constraint = wp.parsed_cond.expr;
-    substitute_result(constraint, ret_val);
-
-    if (wp.action == waypoint::avoid)
-      assume(not2tc(constraint));
-    else
-    {
-      assume(constraint);
-      cur_state->advance_witness_position();
-    }
-    return;
-  }
-}
 
 void goto_symext::symex_step(reachability_treet &art)
 {
@@ -498,36 +408,12 @@ void goto_symext::symex_step(reachability_treet &art)
       }
     }
 
-    // Violation-witness: handle function_enter waypoints in the current segment.
-    // avoid, not matching: skip (persistent).
-    // avoid, matching: skip the call (pc++), segment does not advance.
-    // follow, not matching: stop (follows are ordered; cannot bypass).
-    // follow, matching: advance to next segment; fall through to the call.
-    if (validate_witness && cur_state->cur_seg < cur_state->witness_segs.size())
+    if (validate_witness)
     {
-      const auto &seg = cur_state->witness_segs[cur_state->cur_seg];
-      const auto &loc = cur_state->source.pc->location;
-      for (size_t wp_idx = 0; wp_idx < seg.size(); ++wp_idx)
-      {
-        const waypoint &wp = seg[wp_idx];
-        if (wp.type != waypoint::function_enter)
-          continue;
-        const bool loc_matches =
-          !wp.line_id.empty() && wp.line_id == loc.get_line() &&
-          (wp.function_id.empty() || wp.function_id == loc.get_function());
-        if (loc_matches)
-        {
-          if (wp.action == waypoint::avoid)
-          {
-            cur_state->source.pc++;
-            return;
-          }
-          cur_state->advance_witness_position();
-          break;
-        }
-        if (wp.action != waypoint::avoid)
-          break;
-      }
+      const irep_idt call_line =
+        cur_state->source.pc->location.get_line();
+      if (symex_witness_function_enter(call_line))
+        return;
     }
 
     symex_function_call(deref_code);
